refactoring and statefulset
