--- charts/polkadot-watcher-csv-exporter/templates/persistentVolumeClaim.yaml
@@ -1,23 +1,0 @@
-{{ if ne .Values.environment "ci" }}  
-kind: PersistentVolumeClaim
-apiVersion: v1
-metadata:
-  name: {{ .Release.Name }}-data-csv
-spec:
-  accessModes:
-    - ReadWriteOnce
-  resources:
-    requests:
-      storage: 256Mi
----
-kind: PersistentVolumeClaim
-apiVersion: v1
-metadata:
-  name: {{ .Release.Name }}-data-scanner
-spec:
-  accessModes:
-    - ReadWriteOnce
-  resources:
-    requests:
-      storage: 256Mi
-{{ end }}
